Cast numpy to float32 for test on pooling layers

+++ jacobinet/layers/pooling/average_pooling1d.py
@@ -56,7 +56,7 @@ class BackwardAveragePooling1D(BackwardLinearLayer):
             trainable=False,
         )
         kernel_ = np.ones(pool_size + [1, 1], dtype="float32") / np.prod(pool_size)
-        layer_t.kernel = keras.Variable(kernel_)
+        layer_t.kernel = keras.Variable(kernel_.astype("float32"))
         layer_t.built = True
 
         # shape of transposed input
